gemini-lchain/src/requirements.txt
google-generativeai
tavily-python